extract method to lib
